CAD file of the top frame
ANET changed the top frame, so the one in the assembly video is different.
I created this CAD file which can be e.g. laser-cut out of acrylic or wood. I used the Trotec laser cutter we have in our maker space and it fits just perfectly on top of the printer. Just export is as SVG to use it with such tools.